requirements - Copy.txt
addressable
aiohttp
apiclient
appdirs
#app-store-scraper
#asgiref
asgiref
async-timeout
attrs
backcall
bayesian-optimization
beautifulsoup4
boto3
botocore
bs4
cachetools
certifi
cffi
chardet
click
colorama
configparser
crayons
cssselect
cycler
decorator
defusedxml
#Django
Django==3.2.8
django-allauth
django-cors-headers
django-filter
django-modelclone
django-rest-auth
djangorestframework
django-searchable-select
django-ses
django-sortedm2m
django-sslserver
django-user-agents
docopt
emoji
fake-useragent
feedfinder2
feedparser
future
geographiclib
geoip2
geopy
#googleanalytics
googleanalytics
google-api-core
google-api-python-client
googleapis-common-protos
google-auth
google-auth-httplib2
google-auth-oauthlib
google-play-scraper
httplib2
idna
importlib-metadata
inspect-it
install
ipython
ipython-genutils
jedi
jieba3k
Jinja2
jmespath
joblib
keyring
kiwisolver
lxml
Markdown
MarkupSafe
matplotlib
maxminddb
multidict
mysqlclient
newspaper3k
nltk
num2words
numpy
oauth2client
oauthlib
pandas==1.3.5
parse
parso
pickleshare
#Pillow
Pillow
pip
prettytable
prompt-toolkit
protobuf
pyasn1
pyasn1-modules
pycparser
pycryptodome
pyee
Pygments
pylev
pyparsing
PypeR
pyppeteer
pyquery
pyscape-client
python3-openid
python-dateutil
#python-memcached
pytz
PyYAML
regex
#requests
requests
requests-file
requests-html
requests-oauthlib
rpy2
rsa
s3transfer
#scikit-learn
scikit-learn==1.0.2
#scipy
scipy
seaborn
selenium
setuptools
sgmllib3k
SimilarWeb-Python
six
snakify
soupsieve
sqlparse
threadpoolctl
tinysegmenter
tldextract
tqdm
traitlets
typing-extensions
tzlocal
ua-parser
unicode-slugify
Unidecode
uritemplate
urllib3
user-agents
w3lib
wcwidth
webdriver-manager
websockets
Werkzeug
#xgboost
xgboost==1.6.0
yarl
zipp
pkcs7
pymemcache
django-admin-logs
django-session-security
django-pymemcache
phonenumbers
django-two-factor-auth
twilio
django-formtools
django-password-validators
django-axes
contentful
category_encoders==2.4.0
us
python-memcached
openai